Don't attempt to list all buckets in s3 tests
The tests used a ListBuckets operation which requires privileges that
I'm not comfortable giving to Travis, and which aren't actually
necessary.  We know what the bucket name is so we don't need to scan
each and every bucket, we can just open the one we want.

Also removed an if/then on test bucket creation which caused race
conditions.